Isolate sdk checking logic; Block dnd when sdk below version 0.30
