+++ src/visual_editor/lib/hooks/useFixedPositioning.ts
