+++ src/visual_editor/lib/hooks/useFloatingAnchor.ts
